init checkin
